More moving and fixing
